Changed vars to floats to mitigate value truncation when having small resolution encoders

+++ my_robot_hardware/include/my_robot_hardware/arduino_comms.h
@@ -19,7 +19,7 @@ public:
              int32_t timeout_ms);
   std::string sendEmptyMsg();
   std::string readEncoderValues(int &val_1, int &val_2);
-  std::stringstream setMotorValues(int val_1, int val_2);
+  std::stringstream setMotorValues(double val_1, double val_2);
   std::stringstream setPidValues(float k_p, float k_d, float k_i, float k_o);
   std::string readSerial();
   bool connected() const { return serial_conn_.isOpen(); }
